pyproject.toml
[project]
name = "cdk-pyproject"
version = "0.2.0"
description = "Add your description here"
authors = [{ name = "Masaki Kobayashi", email = "bayashi.cl@gmail.com" }]
dependencies = ["aws-cdk-lib>=2.138.0", "pyproject-metadata>=0.8.0"]
readme = "README.md"
requires-python = ">= 3.11"

classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Framework :: AWS CDK",
    "Framework :: AWS CDK :: 2",
]

[project.urls]
"Source" = "https://github.com/bayashi-cl/cdk-pyproject/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "tox-rye @ git+https://github.com/bluss/tox-rye",
    "virtualenv-rye-discovery @ git+https://github.com/bluss/virtualenv-rye-discovery@0.3.0",
    "tox>=4.14.2",
    "pytest>=8.1.1",
    "mypy>=1.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cdk_pyproject"]

[tool.hatch.version]
path = "src/cdk_pyproject/__init__.py"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN101", "ANN102"]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::SyntaxWarning"]
